+++ packages/npm/skiffa-generator/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./iterable-text-template.js";
 export * from "./iterable.js";
-export * from "./mockable.js";
 export * from "./name.js";
 export * from "./package.js";
